+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jwplayer-capacitor",
-  "version": "0.3.8",
+  "version": "0.4.1",
   "description": "Integration of JWPlayer Web and Mobile SDKs with Capacitor",
   "main": "dist/plugin.cjs.js",
   "module": "dist/esm/index.js",
